done ex 4.21 - 4.23

+++ part-4/controllers/blogs.js
@@ -2,14 +2,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-
-// const getFromtoken = request => {
-//      const authorization = request.get('authorization')
-//      if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-//           return authorization.substring(7)
-//      }
-//      return null
-// }
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
      const blogs = await Blog
@@ -21,16 +14,17 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response, next) => {
      const body = request.body
 
-     console.log("request.token: ", request.token)
-     const decodedToken = jwt.verify(request.token, process.env.SECRET)
+     // console.log("request.token: ", request.token)
+     // const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-     if (!decodedToken.id) {
-          return response.status(401).json({ error: 'token missing or invalid in blogs' })
-     }
+     // if (!decodedToken.id) {
+     //      return response.status(401).json({ error: 'token missing or invalid in blogs' })
+     // }
      if (body.title === undefined || body.url === undefined) {
           return response.status(400).json({ error: 'title or url is missing' })
      }
-     const user = await User.findById(decodedToken.id)
+     console.log("request.user.id: ", request.user.id)
+     const user = await User.findById(request.user.id)
      const blog = new Blog({
           title: body.title,
           author: body.author,
@@ -55,9 +49,22 @@ blogsRouter.get('/:id', async (request, response) => {
      }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-     await Blog.findByIdAndDelete(request.params.id)
-     response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+     // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+     const userToCompare = await User.findById(request.user.id)
+     const blogToDelete = await Blog.findById(request.params.id)
+     // console.log("userToCompare: ", userToCompare._id.toString())
+     // console.log("blogToDelete: ", blogToDelete.user._id.toString())
+     if(userToCompare._id.toString() === blogToDelete.user._id.toString()){
+          try{
+               await Blog.findByIdAndDelete(request.params.id)
+               response.status(204).end()
+          }catch(exception){
+               next(exception)
+          }
+     }else{
+          return response.status(401).json({ error: 'unauthorized' })
+     }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
